+++ autoform-inputs.js
@@ -75,9 +75,17 @@ defaultInputValueHandlers = {
 	'input[type=datetime-local]': function () {
 		var val = this.val();
 		val = (typeof val === "string") ? val.replace(/ /g, "T") : val;
-		var offset = this.attr("data-offset") || "Z";
 		if (Utility.isValidNormalizedLocalDateAndTimeString(val)) {
-			return new Date(val + offset);
+			var timezoneId = this.attr("data-timezone-id");
+			// default is local, but if there's a timezoneId, we use that
+			if (typeof timezoneId === "string") {
+				if (typeof moment.tz !== "function") {
+	        throw new Error("If you specify a timezoneId, make sure that you've added a moment-timezone package to your app");
+	      }
+	      return moment.tz(val, timezoneId).toDate();
+			} else {
+				return new Date(val);
+			}
 		} else {
 			return null;
 		}