+++ package.js
@@ -17,6 +17,7 @@ Package.on_use(function(api) {
     // client
     api.use(['livedata', 'underscore', 'deps', 'templating', 'ui', 'blaze'], 'client');
     api.use('mrt:moment@2.6.0', 'client');
+    api.use('mrt:moment-timezone@0.2.1', 'client', {weak: true});
     api.use(['aldeed:collection2@2.0.0', 'reload'], 'client', {weak: true});
     // Imply SS to make sure SimpleSchema object is available to app
     api.imply('aldeed:simple-schema');